ai-runtime/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-socketio==5.11.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiomysql==0.2.0
psycopg2-binary==2.9.9
pgvector==0.2.4
langgraph>=0.0.26
langgraph-checkpoint-postgres>=0.0.6
langchain>=0.1.10
langchain-openai>=0.0.6
langchain-anthropic>=0.1.4
openai>=1.12.0
anthropic>=0.18.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
redis==5.0.1
sqlparse==0.4.4
python-multipart==0.0.9
aiofiles==23.2.1
tenacity==8.2.3
structlog==24.1.0
nltk==3.8.1
pandas==2.2.0
openpyxl==3.1.2
sqlfluff==3.0.0
sqlglot>=20.0.0
